fix: simulation ends correctly when there are hard faults
* add counter for in-service elevators
* dispatcher periodically notifies scheduler that simulation has ended

Scheduler enters a state of ennui when the last active elevator has a hard fault.
We need to periodically notify it so that it remembers that life is worth living
even if it has no events in the queue! If we don't do this, then he just sits
there waiting for something that will never come.