feat: monitor and display statistics and gracefully end simulation (#101)
* Scratchwork for BK. Has not been run.

* Got code running.

A cute message.

* refactor: change event name to endSimulationEvent

* fix: naming convention for EndSimulationEvent

* feat(Scheduler): add and set flag to end simulation

End it all.

* test: add scenario to test ending simulation without hard faults

* fix: update test scenario 12 description

* feat: add check to end simulation if simulation should end

* refactor: system checks if simulation is over

* feat(FinalState): correctly transition to final state once simulation is over

* feat: display simulation statistics to console at end of simulation

* feat: add recording and displaying of total simulation time

Simulation time is recorded in milliseconds but displayed to user
in HH:MM:SS format. Simulation time is started when Scheduler receives
a StartSimulationEvent.

* fix: simulation ends correctly when there are hard faults

* add counter for in-service elevators
* dispatcher periodically notifies scheduler that simulation has ended

Scheduler enters a state of ennui when the last active elevator has a hard fault.
We need to periodically notify it so that it remembers that life is worth living
even if it has no events in the queue! If we don't do this, then he just sits
there waiting for something that will never come.

* docs(README): add section on total simulation time

* fix: simulation now starts when first floor request is sent

Before, the simulation was starting at time=0, which is wrong.
Simulation should start when first floor request is sent.

* refactor: rename attr from totalElevatorMoveCommandsSent to totalElevatorMovements

* refactor: remove debug statement

Sometimes, less is more.

* style: change attribute header

* Add a drop of clarity to cuteness.

---------

Co-authored-by: Michael De Santis <michaelpeterdesantis@gmail.com>